requirements.txt
colorama==0.4.5
eventlet==0.33.1
Flask==2.1.2
Flask_CKEditor==0.4.6
Flask_Login==0.6.2
flask_paginate==2022.1.8
Flask_SocketIO==5.2.0
Flask_SQLAlchemy==2.5.1
python_engineio==4.3.3
selenium==4.3.0
